Platform separation
Everything I2C and SPI related has been moved from arducam.c to
arducam_arch_raspberrypi.c making porting to new platforms easier.
