+++ ArduCAM4Pi/arducam.c
@@ -61,6 +61,7 @@
 --------------------------------------*/
 
 #include "arducam.h"
+#include "arducam_arch.h"
 #include "utft_spi.h"
 #include "memorysaver.h"
 #include <string.h>
@@ -69,16 +70,6 @@
 #include <time.h>
 #include <linux/i2c-dev.h>
 #include <wiringPiSPI.h>
-
-uint8_t wrSensorReg8_8(uint8_t regID, uint8_t regDat);
-uint8_t rdSensorReg8_8(uint8_t regID, uint8_t* regDat);
-uint8_t wrSensorReg8_16(uint8_t regID, uint16_t regDat);
-uint8_t rdSensorReg8_16(uint8_t regID, uint16_t* regDat);
-uint8_t wrSensorReg16_8(uint16_t regID, uint8_t regDat);
-uint8_t rdSensorReg16_8(uint16_t regID, uint8_t* regDat);
-int wrSensorRegs8_8(const struct sensor_reg reglist[]);
-int wrSensorRegs8_16(const struct sensor_reg reglist[]);
-int wrSensorRegs16_8(const struct sensor_reg reglist[]);
 
 void delayms(int i)
 {
@@ -115,18 +106,15 @@ int PiCAM(uint8_t model)
 			myCAM.sensor_addr = 0x21;
 			break;
 	}
-	spi1 = wiringPiSPISetup (SPI_ARDUCAM, SPI_ARDUCAM_SPEED);
-	if(spi1==-1)
-	{
-		printf("SPI初始化失败！\n");
+	if (!arducam_spi_init()) {
+		printf("ERROR: SPI init failed\n");
 		return 0;
 	}
+
 	// initialize i2c:
-	FD = wiringPiI2CSetup (myCAM.sensor_addr);
-	if(FD==-1)
-	{
+	if (!arducam_i2c_init(myCAM.sensor_addr)) {
+		printf("ERROR: I2C init failed\n");
 		return 0;
-		printf("I2C初始化失败！\n");
 	}
 	return 1;
 }
@@ -297,152 +285,6 @@ uint8_t read_reg(uint8_t addr)
 void write_reg(uint8_t addr, uint8_t data)
 {
 	bus_write(addr | 0x80, data);
-}
-
-void bus_write(uint8_t address, uint8_t value)
-{
-	uint8_t spiData [2] ;
-	spiData [0] = address ;
-  	spiData [1] = value ;
-	wiringPiSPIDataRW (SPI_ARDUCAM, spiData, 2) ;
-}
-
-uint8_t bus_read(uint8_t address)
-{
-	uint8_t spiData[2];
-	spiData[0] = address ;
-	spiData[1] = 0x00 ;
-  	wiringPiSPIDataRW (SPI_ARDUCAM, spiData, 2) ;
-  	return spiData[1];
-}
-
-uint8_t wrSensorReg8_8(uint8_t regID, uint8_t regDat)
-{
-	if(FD != -1)
-	{
-		wiringPiI2CWriteReg8(FD,regID,regDat);
-		return(1);
-	}
-	return 0;
-}
-
-uint8_t rdSensorReg8_8(uint8_t regID, uint8_t* regDat)
-{
-
-	if(FD != -1)
-	{
-		*regDat = wiringPiI2CReadReg8(FD,regID);
-		return(1);
-	}
-	return 0;
-}
-
-uint8_t wrSensorReg8_16(uint8_t regID, uint16_t regDat)
-{
-	if(FD != -1)
-	{
-		wiringPiI2CWriteReg16(FD,regID,regDat);
-		return(1);
-	}
-	return 0;
-}
-
-uint8_t rdSensorReg8_16(uint8_t regID, uint16_t* regDat)
-{
-	if(FD != -1)
-	{
-		*regDat = wiringPiI2CReadReg16(FD,regID);
-		return(1);
-	}
-	return 0;
-}
-
-uint8_t wrSensorReg16_8(uint16_t regID, uint8_t regDat)
-{
-	uint8_t reg_H,reg_L;
-	uint16_t value;
-	reg_H = (regID >> 8) & 0x00ff;
-	reg_L = regID & 0x00ff;
-	value =  regDat << 8 | reg_L;
-	if(FD != -1)
-	{
-		i2c_smbus_write_word_data(FD, reg_H, value);
-		return(1);
-	}
-	return 0;
-}
-
-uint8_t rdSensorReg16_8(uint16_t regID, uint8_t* regDat)
-{
-	uint8_t reg_H,reg_L;
-	int r;
-	reg_H = (regID >> 8) & 0x00ff;
-	reg_L = regID & 0x00ff;
-	if(FD != -1)
-	{
-		r = i2c_smbus_write_byte_data(FD,reg_H,reg_L);
-		if(r<0)
-			return 0;
-		*regDat = i2c_smbus_read_byte(FD);
-		return(1);
-	}
-	return 0;
-}
-
-int wrSensorRegs8_8(const struct sensor_reg reglist[])
-{
-	uint16_t reg_addr = 0;
-	uint16_t reg_val = 0;
-	const struct sensor_reg *next = reglist;
-
-	while ((reg_addr != 0xff) | (reg_val != 0xff))
-	{
-		reg_addr = pgm_read_word(&next->reg);
-		reg_val = pgm_read_word(&next->val);
-		if (!wrSensorReg8_8(reg_addr, reg_val)) {
-			return 0;
-		}
-	   	next++;
-	}
-
-	return 1;
-}
-
-
-int wrSensorRegs8_16(const struct sensor_reg reglist[])
-{
-	unsigned int reg_addr,reg_val;
-	const struct sensor_reg *next = reglist;
-
-	while ((reg_addr != 0xff) | (reg_val != 0xffff))
-	{
-		reg_addr = pgm_read_word(&next->reg);
-		reg_val = pgm_read_word(&next->val);
-		if (!wrSensorReg8_16(reg_addr, reg_val)) {
-			return 0;
-		}
-	   	next++;
-	}
-
-	return 1;
-}
-
-int wrSensorRegs16_8(const struct sensor_reg reglist[])
-{
-	unsigned int reg_addr,reg_val;
-	const struct sensor_reg *next = reglist;
-
-	while ((reg_addr != 0xffff) | (reg_val != 0xff))
-	{
-		reg_addr = pgm_read_word(&next->reg);
-		reg_val = pgm_read_word(&next->val);
-		if (!wrSensorReg8_16(reg_addr, reg_val)) {
-			return 0;
-		}
-	   	next++;
-	}
-
-	return 1;
 }
 
 void OV2640_set_JPEG_size(uint8_t size)
